More cleanup, moved functions to function folder
