+++ pages/functions/errColor.ts
