+++ tsconfig.json
@@ -18,6 +18,6 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "tooltip.js"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "pages/functions/tooltip.js"],
   "exclude": ["node_modules"]
 }
